Hackathon2-Phase4/backend/requirements.txt
fastapi>=0.128.0
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.31
python-multipart>=0.0.6
cryptography>=46.0.3
openai-agents>=0.1.0
mcp>=0.1.0
pydantic>=2.10,<3
pydantic-settings>=2.12.0
pyjwt>=2.10.1
python-dotenv>=1.2.1
slowapi>=0.1.9
sqlalchemy>=2.0.45
passlib[argon2]>=1.7.4
httpx>=0.28.1
psycopg2-binary>=2.9.11
argon2-cffi>=23.1.0
annotated-doc>=0.0.4
anyio>=4.12.0
email-validator>=2.2.0